ref: rename tables files
