+++ src/lib/table/components/table-component/style.ts
